pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ergonet"
version = "1.0.0"
description = "This package allows you to predict the human pose and shape from images"
authors = [
    { name = "Guoyang Zhou", email = "zgy@amazon.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "accelerate==1.9.0",
    "annotated-types==0.7.0",
    "anywidget==0.9.18",
    "asttokens==3.0.0",
    "bbox_visualizer==0.2.2",
    "beautifulsoup4==4.13.4",
    "build==1.2.2.post1",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "chumpy==0.70",
    "comm==0.2.3",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "Cython==3.1.2",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "einops==0.8.1",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fairscale==0.4.13",
    "ffmpeg-python==0.2.0",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "filterpy==1.4.5",
    "fonttools==4.59.0",
    "freetype-py==2.5.1",
    "fsspec==2025.7.0",
    "ftfy==6.3.1",
    "future==1.0.0",
    "gdown==5.2.0",
    "hf-xet==1.1.5",
    "huggingface-hub==0.34.3",
    "idna==3.7",
    "imageio==2.37.0",
    "ipython==8.37.0",
    "ipywidgets==8.1.7",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "joblib==1.5.1",
    "jupyter_bbox_widget==0.6.0",
    "jupyterlab_widgets==3.0.15",
    "kiwisolver==1.4.8",
    "lazy_loader==0.4",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.5",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "ninja==1.11.1.4",
    "numpy==2.2.6",
    "open_clip_torch==3.0.0",
    "opencv-contrib-python==4.12.0.88",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.10.0.84",
    "packaging==25.0",
    "pandas==2.3.1",
    "parso==0.8.4",
    "peft==0.17.0",
    "pexpect==4.9.0",
    "pi_heif==1.1.0",
    "pillow==11.3.0",
    "pillow-avif-plugin==1.5.2",
    "polygraphy==0.49.26",
    "prompt_toolkit==3.0.51",
    "psutil==7.0.0",
    "psygnal==0.14.0",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "py-cpuinfo==9.0.0",
    "pycocotools==2.0.10",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "pyglet==2.1.6",
    "Pygments==2.19.2",
    "pylabel==0.1.55",
    "PyOpenGL==3.1.0",
    "pyparsing==3.2.3",
    "pyproject_hooks==1.2.0",
    "pyrender==0.1.45",
    "PySocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rf100vl==1.0.0",
    "rfdetr==1.2.1",
    "roboflow==1.2.3",
    "safetensors==0.5.3",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "six==1.17.0",
    "smplx==0.1.28",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "supervision==0.26.1",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "timm==1.0.19",
    "tokenizers==0.21.4",
    "tomli==2.2.1",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchvision==0.22.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.54.1",
    "trimesh==4.7.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "widgetsnbextension==4.0.14",
    "yacs==0.1.8",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
namespaces = false

[tool.setuptools.package-data]
"ergonet.backbones.hrnet.models" = ["*.yaml"]